move more tests to sliver persistent header test
